# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: SoulAuth
  version: "0.1.0"
  description: >-
    Actor-native identity and authentication infrastructure.
    Endpoint surface only; payload schemas are not yet frozen.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: "{APP_URL}"
    description: >-
      The public address, set by APP_URL. It is also the OIDC issuer — the two
      must match character for character.
    variables:
      APP_URL:
        default: http://localhost:8080
components:
...
  schemas:
    Error:
      type: object
      required:
      - error
      - message
      additionalProperties: true
      description: 'The single error envelope used across the whole API. `error` is a **stable machine-readable
        code** — branch on it. `message` is prose: its wording may improve at any time and must not be
        used for decisions. Individual errors carry documented additional fields.'
      properties:
        error:
          type: string
          description: Stable error code, snake_case.
          enum:
          - internal_error
          - invalid_credentials
          - email_not_verified
          - invalid_token
          - user_not_found
          - email_exists
          - username_exists
          - oauth_error
          - password_already_set
          - invalid_user_id
          - not_found
          - validation_error
          - permission_denied
          - insufficient_permissions
          - missing_permission
          - account_suspended
          - account_inactive
          - account_deleted
          - account_locked
          - rate_limit_exceeded
          - forbidden
          - bad_request
          - not_configured
          - unauthorized
          - service_unavailable
        message:
          type: string
          description: Human-readable explanation. Not part of the contract.
        required_permission:
          type: string
          description: Carried by `missing_permission` only — which permission is missing.
        locked_until_seconds:
          type: integer
          description: Carried by `account_locked` only — seconds until the lock clears.
...
    AccountStatusResponse:
      properties:
        reason:
          type: string
        status:
          type: object
        updated_at:
          type: object
        updated_by:
          type: string
        user_id:
          type: string
      required:
      - user_id
      - status
      - updated_at
      - updated_by
      type: object
...
    ActivityLogResponse:
      properties:
        activities:
          items:
            $ref: '#/components/schemas/UserActivityResponse'
          type: array
        limit:
          type: integer
        page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required:
      - activities
      - total
      - page
      - limit
      - total_pages
      type: object
...
    UpdateAccountStatusRequest:
      properties:
        reason:
          type: string
        status:
          type: object
      required:
      - status
      type: object
    UpdateMembershipRequest:
      properties:
        membership_expiry:
          type: object
        membership_level:
          type: string
      required:
      - membership_level
      type: object
...
    UserActivityResponse:
      properties:
        action:
          type: string
        category:
          type: object
        details:
          type: object
        id:
          type: string
        ip_address:
          type: string
        status:
          type: object
        timestamp:
          type: object
        user_agent:
          type: string
        user_id:
          type: string
      required:
      - id
      - user_id
      - action
      - category
      - ip_address
      - user_agent
      - details
      - status
      - timestamp
      type: object
...
    UserListResponse:
      properties:
        limit:
          type: integer
        page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        users:
          items:
            $ref: '#/components/schemas/UserResponse'
          type: array
      required:
      - users
      - total
      - page
      - limit
      - total_pages
      type: object
    UserPreferencesResponse:
      properties:
        created_at:
          type: object
        date_format:
          type: string
        email_notifications:
          type: boolean
        id:
          type: string
        language:
          type: string
        marketing_emails:
          type: boolean
        newsletter:
          type: boolean
        security_emails:
          type: boolean
        sms_notifications:
          type: boolean
        theme:
          type: string
        time_format:
          type: string
        timezone:
          type: string
        updated_at:
          type: object
        user_id:
          type: string
      required:
      - id
      - user_id
      - theme
      - language
      - email_notifications
      - sms_notifications
      - marketing_emails
      - security_emails
      - newsletter
      - timezone
      - date_format
      - time_format
      - created_at
      - updated_at
      type: object
    UserProfileResponse:
      properties:
        avatar_url:
          type: string
        bio:
          type: string
        created_at:
          type: object
        date_of_birth:
          type: object
        display_name:
          type: string
        first_name:
          type: string
        id:
          type: string
        last_name:
          type: string
        locale:
          type: string
        location:
          type: string
        phone:
          type: string
        timezone:
          type: string
        updated_at:
          type: object
        user_id:
          type: string
        website:
          type: string
      required:
      - id
      - user_id
      - created_at
      - updated_at
      type: object
    UserResponse:
      properties:
        account_status:
          type: object
        created_at:
          type: object
        email:
          type: string
        has_password:
          type: boolean
        id:
          type: string
        is_admin:
          type: boolean
        last_login_at:
          type: object
        membership_expiry:
          type: object
        membership_level:
          type: string
        username:
          type: string
        verified:
          type: boolean
      required:
      - id
      - email
      - username
      - is_admin
      - verified
      - membership_level
      - created_at
      - has_password
      - account_status
      type: object
...
  securitySchemes:
    # 三种，不是一种。
    #
    # 把它们混成一个 `bearerAuth` 会产生具体的误导：读契约的人会以为
    # /api/oidc/userinfo 收的是登录拿到的那枚会话令牌 —— 它收的是 OIDC
    # 访问令牌，两者由不同端点签发、生命周期也不同。
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Session token, from POST /api/auth/login (or POST
        /api/actors/authenticate for an AI actor). Used by every /api business
        endpoint.
    oidcAccessToken:
      type: http
      scheme: bearer
      description: >-
        OIDC access token, from POST /api/oidc/token. Used by
        /api/oidc/userinfo **only**. Not interchangeable with the session token
        above.
    browserSession:
      type: apiKey
      in: cookie
      name: soulauth_session
      description: >-
        Browser session cookie. /api/oidc/authorize uses it to decide whether
        this browser is logged in; its absence is a redirect to the login page,
        not an error.
...
paths:
...
  /api/users:
    get:
      tags:
      - Actors & Profiles
      operationId: user_management_list_users
      description: >-
        List accounts. Requires `soulauth:users.read`.
      description_zh: >-
        列出账户，需要 `soulauth:users.read` 权限。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:users.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
      - name: page
        in: query
        required: false
        schema:
          type: integer
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
      - name: sort_order
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: object
  /api/users/:user_id:
    get:
      tags:
      - Actors & Profiles
      operationId: user_management_get_user_by_id
      description: >-
        One account by id.
      description_zh: >-
        按 id 返回单个账户。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:users.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
  /api/users/:user_id/activity-log:
    get:
      tags:
      - Actors & Profiles
      operationId: user_management_get_user_activity_log_by_id
      description: >-
        Another actor's activity log. Requires `soulauth:audit.read`, not `users.read`.
      description_zh: >-
        返回其他主体的活动记录。需要的是 `soulauth:audit.read` 而不是 `users.read`。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:audit.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityLogResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      - name: category
        in: query
        required: false
        schema: *id008
      - name: end_date
        in: query
        required: false
        schema: *id009
      - name: limit
        in: query
        required: false
        schema: *id010
      - name: page
        in: query
        required: false
        schema: *id011
      - name: start_date
        in: query
        required: false
        schema: *id012
      - name: status
        in: query
        required: false
        schema: *id013
  /api/users/:user_id/membership:
    put:
      tags:
      - Actors & Profiles
      operationId: user_management_update_user_membership
      description: >-
        Change membership level and expiry.
      description_zh: >-
        修改会员等级与到期时间。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:users.write
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMembershipRequest'
  /api/users/:user_id/preferences:
    get:
      tags:
      - Actors & Profiles
      operationId: user_management_get_user_preferences_by_id
      description: >-
        Another actor's preferences.
      description_zh: >-
        返回其他主体的偏好设置。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:users.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
  /api/users/:user_id/profile:
    get:
      tags:
      - Actors & Profiles
      operationId: user_management_get_user_profile_by_id
      description: >-
        Another actor's profile.
      description_zh: >-
        返回其他主体的档案。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:users.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
  /api/users/:user_id/status:
    put:
      tags:
      - Actors & Profiles
      operationId: user_management_update_user_account_status
      description: >-
        Change an account's status. Suspension stops **future** authentication; sessions on
        other replicas persist for up to one cache TTL. History is not rewritten.
      description_zh: >-
        修改账户状态。停用挡住的是**未来**的认证；其它副本上已有的会话最多还能存续一个缓存 TTL。历史不被改写。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:users.write
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStatusResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountStatusRequest'
...